Add NSWay conforming NSObjectProtocol
